(refactor): 코드 리뷰 반영한 리펙토링
